fix rate type

+++ src/app/dto/add.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -12,4 +13,4 @@ class AddCurrencyDTO:
 class AddRateDTO:
     base_currency: str
     target_currency: str
-    rate: float
+    rate: Decimal
